GTE: shift on the unsigned type where the value can be negative
Running the test suites under UBSan -- the build scripts already took a
SANITIZER variable, which nothing had used -- turned up eighteen sites in
IopGte.cpp shifting a negative signed value left, which is undefined in
C++ before C++20.

They are all the same two idioms, and the GTE does both constantly: every
translation vector is scaled by 1<<12 before the accumulate, and the
44-bit accumulator is sign-extended with the usual shift-up-shift-down
pair. A gte_shl helper does the shift on the unsigned type and converts
back, which is bit-identical on two's complement and defined everywhere.

The arithmetic is unchanged, and the console log says so: the fuzz replay
in tests/iop/hwgtefuzz still matches on all 1100 tests, before and after.
That mattered here more than usual -- this is exactly the kind of tidying
that silently changes a saturation edge, and without the oracle from a few
commits ago it would have been an unverifiable change to working code.

Two of the eighteen were missed on the first pass because the pattern
matched tx but not ty and tz on the following lines. UBSan named them
straight away, which is the argument for running it rather than reading
for the pattern.

Also fixes a signed overflow of my own in tests/mmi/hwscore.c: the PMADDH
and PMSUBH accumulate wraps rather than saturating, which is what the
hardware does, so it now wraps on the unsigned type. All five suites are
clean under UBSan.

+++ tests/mmi/hwscore.c
@@ -101,11 +101,15 @@ static void apply(int op, const Q* rs, const Q* rt, Q* HI, Q* LO, Q* rd)
 		{
 			const s16 a = (s16)(rs->UL[i/2] >> ((i & 1) * 16));
 			const s16 b = (s16)(rt->UL[i/2] >> ((i & 1) * 16));
-			s32 p = (s32)a * (s32)b;
+			/* The accumulate wraps rather than saturating, which is
+			 * what the hardware does, so do it on the unsigned type:
+			 * signed overflow here is undefined and the operands do
+			 * reach it -- UBSan on this file reports it otherwise. */
+			u32 p = (u32)((s32)a * (s32)b);
 			Q* d = toLo[i] ? LO : HI;
-			if (op == OP_PMADDH) p = (s32)d->UL[word[i]] + p;
-			else if (op == OP_PMSUBH) p = (s32)d->UL[word[i]] - p;
-			d->UL[word[i]] = (u32)p;
+			if (op == OP_PMADDH) p = d->UL[word[i]] + p;
+			else if (op == OP_PMSUBH) p = d->UL[word[i]] - p;
+			d->UL[word[i]] = p;
 		}
 		rd->UL[0] = LO->UL[0]; rd->UL[1] = HI->UL[0];
 		rd->UL[2] = LO->UL[2]; rd->UL[3] = HI->UL[2];
